0.3

+++ controllers/contacts/removeById.js
@@ -4,11 +4,12 @@ const mongoose = require("mongoose");
 
 const removeById = async (req, res) => {
   const { id } = req.params;
+  const ownerId = req.user._id;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     throw new Error("id is not valid");
   }
-  const data = await Contact.findByIdAndRemove(id);
+  const data = await Contact.findByIdAndRemove({ _id: id, owner: ownerId });
 
   if (!data) {
     throw new NotFound("Not found");
